Refactor UI tabs into separate modules (#12)
* Refactor UI tabs into modules

* Add tab visibility and ordering preferences
