+++ ui_constants.py
@@ -1,0 +1,2 @@
+SETTINGS_ENABLED_TIERS = "enabled_tiers"
+SETTINGS_CRAFT_6X6_UNLOCKED = "crafting_6x6_unlocked"
