+++ ui_tabs/__init__.py
@@ -1,0 +1,1 @@
+"""UI tab modules for the GTNH Helper application."""
